+++ odd_collector/adapters/druid/adapter.py
@@ -1,13 +1,13 @@
 from typing import List, Type
 
 from odd_collector_sdk.domain.adapter import AbstractAdapter
-from odd_models.models import DataEntityList, DataEntity, DataEntityType, DataSet, DataSetField, DataSetFieldType
+from odd_models.models import DataEntityList, DataEntity
 
-from odd_collector.adapters.druid.mappers.client import DruidBaseClient, DruidClient
-from odd_collector.adapters.druid.mappers.column import Column
-from odd_collector.adapters.druid.mappers.generator import DruidGenerator
-from odd_collector.adapters.druid.mappers.table import Table
-from odd_collector.adapters.druid.mappers.types import TYPES_DRUID_TO_ODD
+from odd_collector.adapters.druid.client import DruidBaseClient, DruidClient
+from odd_collector.adapters.druid.domain.column import Column
+from odd_collector.adapters.druid.domain.table import Table
+from odd_collector.adapters.druid.generator import DruidGenerator
+from odd_collector.adapters.druid.mappers.tables import table_to_data_entity
 from odd_collector.domain.plugin import DruidPlugin
 
 
@@ -24,60 +24,27 @@ class Adapter(AbstractAdapter):
     def get_data_source_oddrn(self) -> str:
         return self.__oddrn_generator.get_data_source_oddrn()
 
-    def get_data_entity_list(self) -> DataEntityList:
+    async def get_data_entity_list(self) -> DataEntityList:
         return DataEntityList(
             data_source_oddrn=self.get_data_source_oddrn(),
-            items=(self.get_data_entities()),
+            items=(await self.get_data_entities()),
         )
 
-    def get_data_entities(self) -> List[DataEntity]:
+    async def get_data_entities(self) -> List[DataEntity]:
         # Fetch
-        tables: List[Table] = self.client.get_tables()
-        columns: List[Column] = self.client.get_columns()
-        tables_nr_of_rows: dict = self.client.get_tables_nr_of_rows()
-
-        # Set oddrn
-        [self.set_table_oddrns(table) for table in tables]
-        [self.set_column_oddrns(column) for column in columns]
+        tables: List[Table] = await self.client.get_tables()
+        columns: List[Column] = await self.client.get_columns()
+        tables_nr_of_rows: dict = await self.client.get_tables_nr_of_rows()
 
         # Transform
         data_entities = [
-            self.table_to_data_entity(table,
-                                      list(filter(lambda column: column.table == table.name, columns)),
-                                      tables_nr_of_rows.get(table.name, 0))
-            for table in tables
+            table_to_data_entity(
+                self.__oddrn_generator,
+                table,
+                list(filter(lambda column: column.table == table.name, columns)),
+                tables_nr_of_rows.get(table.name, 0)
+            ) for table in tables
         ]
 
         # Return
         return data_entities
-
-    def table_to_data_entity(self, table: Table, columns: List[Column], nr_of_rows: int) -> DataEntity:
-        # Return
-        return DataEntity(
-            oddrn=self.__oddrn_generator.get_oddrn_by_path("tables", table.name),
-            name=table.name,
-            type=DataEntityType.DATABASE_SERVICE,
-            metadata=[],
-            dataset=DataSet(
-                rows_number=nr_of_rows,
-                field_list=list(map(self.column_to_data_set_field, columns))
-            )
-        )
-
-    def column_to_data_set_field(self, column: Column) -> DataSetField:
-        from odd_models import models
-        return DataSetField(
-            oddrn=self.__oddrn_generator.get_oddrn_by_path("columns", column.name),
-            name=column.name,
-            type=DataSetFieldType(
-                type=TYPES_DRUID_TO_ODD.get(column.type, models.Type.TYPE_UNKNOWN),
-                logical_type=column.type,
-                is_nullable=column.is_nullable
-            )
-        )
-
-    def set_table_oddrns(self, table: Table):
-        self.__oddrn_generator.set_oddrn_paths(**{"catalogs": table.catalog, "schemas": table.schema, "tables": table.name})
-
-    def set_column_oddrns(self, column: Column):
-        self.__oddrn_generator.set_oddrn_paths(**{"catalogs": column.catalog, "schemas": column.schema, "tables": column.table, "columns": column.name})
