+++ odd_collector/adapters/druid/generator.py
@@ -1,7 +1,7 @@
 from oddrn_generator import Generator
 from oddrn_generator.server_models import HostnameModel
 
-from odd_collector.adapters.druid.mappers.path_model import DruidPathsModel
+from odd_collector.adapters.druid.path_model import DruidPathsModel
 
 
 class DruidGenerator(Generator):
