--- odd_collector/adapters/druid/mappers/client.py
@@ -1,165 +1,0 @@
-import json
-from abc import ABC, abstractmethod
-from typing import List
-
-import requests
-from odd_collector_sdk.errors import MappingDataError, DataSourceError
-
-from odd_collector.adapters.druid.mappers.column import Column
-from odd_collector.adapters.druid.mappers.table import Table
-from odd_collector.domain.plugin import DruidPlugin
-
-
-class DruidBaseClient(ABC):
-    @abstractmethod
-    def get_tables(self) -> List[Table]:
-        raise NotImplementedError
-
-    @abstractmethod
-    def get_columns(self) -> List[Column]:
-        raise NotImplementedError
-
-    def get_tables_nr_of_rows(self) -> dict:
-        raise NotImplementedError
-
-
-class DruidClient(DruidBaseClient):
-    def __init__(self, config: DruidPlugin) -> None:
-        self.__config = config
-
-    def get_tables(self) -> List[Table]:
-        """
-        Get tables from druid.
-
-        Note: It excludes the INFORMATION_SCHEMA that keep druid metadata related
-        to tables persisted.
-
-        Note: It excludes system tables related to druid operations.
-
-        :return: a list of tables
-        """
-        # Prepare
-        tables: List[Table] = []
-        url = f"{self.__config.host}:{self.__config.port}/druid/v2/sql/"
-        sql_query = """
-            SELECT 
-              "TABLE_CATALOG", 
-              "TABLE_SCHEMA", 
-              "TABLE_NAME", 
-              "TABLE_TYPE", 
-              "IS_JOINABLE", 
-              "IS_BROADCAST"
-            FROM INFORMATION_SCHEMA.TABLES
-            WHERE LOWER("TABLE_SCHEMA") != 'information_schema'
-              AND LOWER("TABLE_SCHEMA") != 'sys'
-        """
-
-        # Execute
-        try:
-            result = requests.post(url, json={"query": sql_query})
-            result_dict = json.loads(result.text)
-        except Exception as e:
-            # Throw
-            raise DataSourceError(f"Couldn't execute Druid query: {sql_query}") from e
-
-        # Transform
-        try:
-            for row in result_dict:
-                tables.append(Table(
-                    row["TABLE_CATALOG"],
-                    row["TABLE_SCHEMA"],
-                    row["TABLE_NAME"],
-                    row["TABLE_TYPE"],
-                    row["IS_JOINABLE"],
-                    row["IS_BROADCAST"]))
-        except Exception as e:
-            # Throw
-            raise MappingDataError("Couldn't transform Druid result to Table model") from e
-
-        # Return
-        return tables
-
-    def get_columns(self) -> List[Column]:
-        # Prepare
-        table_columns: List[Column] = []
-        url = f"{self.__config.host}:{self.__config.port}/druid/v2/sql/"
-        sql_query = """
-            SELECT
-              TABLE_CATALOG,
-              TABLE_SCHEMA,
-              TABLE_NAME,
-              COLUMN_NAME,
-              ORDINAL_POSITION,
-              COLUMN_DEFAULT,
-              IS_NULLABLE,
-              DATA_TYPE,
-              CHARACTER_MAXIMUM_LENGTH,
-              CHARACTER_OCTET_LENGTH,
-              NUMERIC_PRECISION,
-              NUMERIC_PRECISION_RADIX,
-              NUMERIC_SCALE,
-              DATETIME_PRECISION,
-              CHARACTER_SET_NAME,
-              COLLATION_NAME,
-              JDBC_TYPE
-            FROM INFORMATION_SCHEMA.COLUMNS
-        """
-
-        # Execute
-        try:
-            result = requests.post(url, json={"query": sql_query})
-            result_dict = json.loads(result.text)
-        except Exception as e:
-            # Throw
-            raise DataSourceError(f"Couldn't execute Druid query: {sql_query}") from e
-
-        # Transform
-        try:
-            for row in result_dict:
-                column = Column(row["TABLE_CATALOG"],
-                                row["TABLE_SCHEMA"],
-                                row["TABLE_NAME"],
-                                row["COLUMN_NAME"],
-                                row["DATA_TYPE"],
-                                True if row["IS_NULLABLE"] == 'YES' else False,
-                                row["ORDINAL_POSITION"])
-                table_columns.append(column)
-        except Exception as e:
-            # Throw
-            raise MappingDataError("Couldn't transform Druid result to Column model") from e
-
-        # Return
-        return table_columns
-
-    def get_tables_nr_of_rows(self) -> dict:
-        # Prepare
-        tables_nr_of_rows = {}
-        url = f"{self.__config.host}:{self.__config.port}/druid/v2/sql/"
-        sql_query = f"""
-        SELECT 
-            datasource,
-            CASE WHEN SUM(num_rows) = 0 THEN 0 ELSE SUM(num_rows) / (COUNT(*) FILTER(WHERE num_rows > 0)) END AS avg_num_rows
-        FROM sys.segments
---      todo: is_active shall be considered for newest druid versions   
-        WHERE is_available = 1
-        GROUP BY 1
-        """
-
-        # Execute
-        try:
-            result = requests.post(url, json={"query": sql_query})
-            result_dict = json.loads(result.text)
-        except Exception as e:
-            # Throw
-            raise DataSourceError(f"Couldn't execute Druid query: {sql_query}") from e
-
-        # Transform
-        try:
-            for row in result_dict:
-                tables_nr_of_rows[row["datasource"]] = row["avg_num_rows"]
-        except Exception as e:
-            # Throw
-            raise MappingDataError("Couldn't transform Druid result for number of rows per table") from e
-
-        # Return
-        return tables_nr_of_rows
