--- odd_collector/adapters/druid/mappers/column.py
@@ -1,12 +1,0 @@
-from dataclasses import dataclass
-
-
-@dataclass
-class Column:
-    catalog: str
-    schema: str
-    table: str
-    name: str
-    type: str
-    is_nullable: bool
-    ordinal_position: int
