--- odd_collector/adapters/druid/mappers/table.py
@@ -1,11 +1,0 @@
-import dataclasses
-
-
-@dataclasses.dataclass
-class Table:
-    catalog: str
-    schema: str
-    name: str
-    type: str
-    is_joinable: str
-    is_broadcast: str
