+++ odd_collector/adapters/druid/path_model.py
